Refactor: move RequestManager to the bot package
Relocated RequestManager from the conversation package to the bot package for improved structure and maintainability. Updated references and imports across the codebase to reflect the new location. No functional changes were introduced.
