+++ pkg/bot/request_manager_test.go
@@ -1,4 +1,4 @@
-package conversation
+package bot
 
 import (
 	"testing"
