+++ pkg/bot/service_test.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
-	"github.com/ksysoev/help-my-pet/pkg/conversation"
 	"github.com/ksysoev/help-my-pet/pkg/core"
 	"github.com/ksysoev/help-my-pet/pkg/i18n"
 	"github.com/stretchr/testify/assert"
@@ -154,7 +153,7 @@ func TestServiceImpl_ProcessMessage(t *testing.T) {
 				Bot:      mockBot,
 				AISvc:    mockAI,
 				Messages: &i18n.Config{},
-				reqMgr:   conversation.NewRequestManager(),
+				reqMgr:   NewRequestManager(),
 			}
 
 			tt.setupMocks(mockBot, mockAI)
@@ -176,7 +175,7 @@ func TestServiceImpl_Run(t *testing.T) {
 		Bot:      mockBot,
 		AISvc:    mockAI,
 		Messages: &i18n.Config{},
-		reqMgr:   conversation.NewRequestManager(),
+		reqMgr:   NewRequestManager(),
 	}
 
 	updates := make(chan tgbotapi.Update)
